Fixed Animation and Physics bug Added BulletObject with Animation
Animations and Physics are now in FixedUpdate so it now the game will run at 60 fps even on 60+ hz monitors. Bullet Object has been made with Simple Animation and will spawn on the shooting command, a.k.a. on the click of the mouse. Currently, there is only one Bullet object